scrapper/requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
